Fix build for architectures with broken seccomp
