+++ Cargo.toml
@@ -36,3 +36,4 @@ toml = "0.4"
 serde = "1.0"
 serde_derive = "1.0"
 users = "0.6"
+cfg-if = "0.1"
